Pull out individual components.
